Started work on reusable widgets
Started work on two reusable widgets: a tile-based scrollbar and a keybind strip. These are not yet used or tested.

(We already have a reusable scrollbar, but it paints into a Window, which means it consumes several more BG tiles than we really need. If we commit to requiring scrollbars to be placed within a single BG tile, then we can do scrollbars of arbitrary height using no more than 4 BG tile IDs.)

Updated the repo readme to no longer be a copy of the readme for the savedata system.