Tile-based scrollbar is now done
Refactored scrollbar initialization: the parameters needed for init are now in their own struct, so you can use designated initializers. (C doesn't have struct-assignment, so you can only use initializer syntax to initialize, not to overwrite.) This also means scrollbars are initialized similarly to keybind strips.

Fixed implementation bugs in the tile-based scrollbar code. This includes incorrect argument ordering for FillTiles calls, incorrect comparisons in the edge-drawing code, a failure to guard against the case of there being zero scrollable items, a failure to compute the thumb position properly (evident when scrolled to the bottom of a pane), and some bad tile Y-coordinates.

The Custom Game Options menu now uses a tile-based scrollbar, instead of the old pixel-based scrollbar. This means we should be using fewer BG tiles for rendering the options listing now.

Added a couple of sanity assertions to the keybind strip code.