Sped up time-to-first-interaction
Made the following changes to speed up the game's bootup.

Halved the duration of the fade-in from white. Halved the amount of time for which the copyright screen is shown. Removed a fade effect that made the transition from the copyright screen to the intro cutscene feel sluggish.

Additionally, documented some of the GameCube multi-boot logistics here.