Ported over coalesced item-pickup text
I made this change in my original exploratory fork/branch, and forgot to port it over to this new repo: item pickup text is now coalesced. That is: we now use one dialog box to tell you that you've picked up an item, and what bag pocket you've put it in, rather than splitting that into two dialog boxes.