Made it possible to adjust the letterbox timings in the battle intro
Figured out how the receding letterbox effect in the battle intro works, and made it possible to adjust the timings, albeit with some constraints (a more thorough edit of the code involved would be needed to allow more freeform speeds). This doesn't make the battle UI responsive faster, because the ultimate constraint is the slide-in animation speed, but it does make the UI *feel* slightly faster.

I'm not increasing the slide-in speed beyond 50%, in part because doubling it makes it feel too fast, and because if I double it without adjusting the timings on the parallax effects (which I haven't touched), then the animation just looks wrong. (And in turn, making the parallax effects scroll by too quickly might make them hard to "read.")