LuEnumPicker; started work on species selection menu
The Custom Game Options menu applies a visual effect to any option that the cursor is over, to show arrows to either side and animate them as the option's value is changed by the user. This has been moved to its own widget, LuEnumPicker, so that we don't need as much boilerplate and raw graphics definitions in menu code.

Started work on a submenu for picking a species. Not even remotely ready to test yet.

Added a roadmap file, and updated the changelog.