PokeCenter nurse animation change
Changed the PokeCenter healing animation so that the nurse takes Poke Balls from the player two at a time, and inserts them two at a time. The implementation is incredibly janky; ditching the dummy sprite used to spawn Poke Ball sprites, and just doing everything in the task, would be more manageable.