Repels can be reused from the overworld now
When a repel runs out, you're now prompted to use another, if you have one. The notification also shows the repel type that ran out.

To facilitate this feature, using a repel now stores the item type used (we've co-opted unused var 0x404E), and we have a script command that returns how many of an item you have.