The player can now voluntarily forfeit trainer battles
The player can now voluntarily forfeit trainer battles. This functionality already existed in the vanilla game, but only for Battle Frontier and Trainer Hill battles. The relevant checks have been moved to a helper function, and the battle scripts and core have been modified so that forfeiting battles outside of those contexts counts as a loss.

Additionally, the player's battle controller has been updated so that in battles where forfeiting is possible, the "RUN" option is renamed to "GIVE UP."