Battle controllers reset their latent state; other work
Unified the behavior of <Controller>BufferExecCompleted.

Battle controllers now clear their per-controller latent state on startup and on shutdown. This means that the latent state storage is guaranteed to be zeroed when the controller first receives and starts handling a message, such that some latent tasks can now be handled "top-level" without needing to alter the callback.

We no longer use a separate battle controller message for playing an attack string and an attack animation in unison. Instead, we've repurposed the high bit of one of the "move animation" message parameters, to serve as an indicator for whether to also play an attack string.

Added a battle script command to both show an attack string and play a move animation, in unison. All that's left is to edit the existing battle scripts to use this... which will be a considerable challenge, since it's all threaded code with tangled jumps and interdependencies.