Started work on a unified script command for stat changes
There are two different sets of script commands used to enact stat changes during battle, with lots of functionality split between them. In order to make stat change text and animations play in unison, we need to unify the approaches used to enact stat changes.

This commit adds `trystatchange`, a script command which attempts to perform a stat change and then reports the result via a new battle controller message. The message handler dynamically constructs strings to report the stat changes (TODO: dynamically place newlines), and plays the animation in unison.

We also add some battle string formatting codes for printing the name or ability of a specific battler by index, just to make things a bit easier on ourselves on the UI end of things, by not having to track what the hazily-defined-and-delineated contextual battler IDs hold.