Work on VUI naming screen
VUI: Made it so that widgets with their own callbacks accept those callbacks at init time. Made it so that VUIKeyboard accepts its destination string-buffer-ref at init time.

The VUI naming screen has some more graphics, as I continue to work on its layout. (The backdrop is a recolor of the "Crag" PC box background.) We're currently in a transitional phase of sorts, between "keyboard on the left; buttons on the right" and "keyboard on the left; OK/Backspace on the right; dedicated per-charset buttons in a row below."

Added AGB_WARNINGs to Game Freak's sprite library, in CreateSpriteAt, to validate input parameters.

There are some aspects of Game Freak's BG library that suck, frankly. Wrote up some plans to improve that, as part of the to-do list for the VUI naming screen. I've been bitten by these issues more than once and frankly, I'm sick of it.