VUI naming screen work
VUI-based naming screen now uses a custom-keyboard widget (and so defines keyboard pages for accented letters) and supports title text, which can be set by the caller. Removed the "next charset" button, but kept the keybind to void interfering with users' muscle memory.

Fixed mistakes in the menu teardown code.

VUI-based naming screen now uses a more advanced format for the charset buttons. Each button has a background and outline color done up in its own color within the palette. We set all such colors to white by default, and this means we can show an outline around the button the cursor is over, or shade in the currently active button, by just messing with the palette.