Update field_debug_menu.c
Added a handler for Secret Power. Untested.

Adjusted how the handler for Flash works, in an attempt to fix a bug that occurs when using Flash on a multi-map cave. Basically, you'd use Flash successfully, but then, when returning from any non-overworld screen or traveling between multiple maps in the same cave, the game would try to use Flash again, softlocking with a glitchy flash shape. Fix untested.