VUI naming screen work
Letters are centered more accurately within the blanks in a VUI keyboard value, and the grid cells in a VUI custom keyboard.

Redesigned the naming screen background: it is now a simple drawing of an emerald jewel.

Added borders to the naming screen's keyboard and keyboard value widgets.

The naming screen's default title for renaming a Pokemon now matches vanilla.

The naming screen now fades to black on exit, so there isn't a one-frame glimpse of the torn-down background layers.

Fixed a use-after-free at menu exit.

The VUI naming screen is now used when nicknaming Pokemon via the party menu, though not when nicknaming them on capture or via the Name Rater.

Tried and failed to improve how the VUI custom-keyboard widget positions key columns within its width. There's extra space at its right edge that it isn't using for some reason. Code and definitions relating to keyboard metrics are messy anyway, and should probably be redone (perhaps even to the point of allowing the keyboard width in tiles to vary per widget).