VUI header fixes; item_menu.c organization
Fixed header include order for VUI headers.

Wrote custom Yes/No dialog implementations, but I don't know if I want to use them yet.

Added code-folding regions to `item_menu.c`. Notepad++ (or rather, the lexer project it relies on) doesn't support `#pragma region`, so I have to use a cheap hack for now (`#if 1`).

I'm trying to find ways to improve how the bag menu code is organized, and ways to split it into smaller, more manageable parts. One of the things I want to do is make it possible to toggle the menu between the vanilla list view, and a grid view of item icons and quantities (think Ocarina of Time, basically), as the latter would allow faster navigation across a single bag pocket during situations like using items in battles. Changes on that scale are a bit challenging when 70% of the menu implementation is just dumpied into one giant file.