Walk-through-walls stability fix
The "walk through walls" debug command triggered edge-cases within the vanilla map connection logic, such that when walking outside of a map's boundaries in some cases, you could trick the game into following a nullptr map connection. Functionally, this had the effect of corrupting the on-screen tilemap, and instantly placing the player in a glitched map (identified as Petalburg City, with Gym music, and functioning as an interior so you can't Fly away). Based on mGBA's logs, it also caused the game to overflow a buffer somewhere (probably when backing up the map view as part of the map transition).

This bugfix remedies that issue.

However, there remain bugs with map connections that can be caused by walk-through-walls. For a simple example, travel to Fallarbor Town and then exit the town to the west. Travel south a bit, until you're below Fallarbor Town's southern edge. Then, travel east, so that you're out of bounds south of Fallarbor Town, and then travel north to enter Fallarbor Town from the south. The on-screen tiles will be shifted around, and border tiles will be drawn in-bounds, because the "map view" buffer is not properly updated -- maybe because of your movement direction (i.e. following a west-to-east map connection by moving north), maybe because of your position (i.e. following a map connection without physically walking over the connected edge), or maybe both. I'm not really sure how to fix that yet. Still, it's better than the previous state of affairs.