lu-save-js: Critical fix to bitstreams
Fixed a major error that could occur when trying to write values above 0x7FFFFFFF to a bitstream, wherein the values would be improperly sign-extended instead of zero-extended when trying to extract slices of their bits. This was corrupting certain values in the bitstream based on what values were serialized after them, with Pokemon checksums being at particular risk (i.e. if one Pokemon's personality value was sufficiently high, it'd corrupt the checksum of the previous Pokemon).

Added a display override for SerializedBoxPokemon, to call attention to Bad Eggs.

Fixed CBufferifier's handling of booleans.