Trainer Card startup optimization complete
Took states that were already combined via fall-through, and "formally" combined them, reducing the number of states listed in the respective enums. This will make it easier to try and accurately count how many frames we're burning on startup. Also took this opportunity to combine a couple more states.

Frame counts:
 - 4 main-state frames
 - 4 init-state frames
 - 3 gfx-state frames (4 states; 1 state overlaps with init-state)
 - 11 total frames

With these changes, the Trainer Card opens as fast as the Bag, now.