Incomplete menu refactor; ROM does not yet build
Trying to convert the new naming screen into a "short string entry menu," so named because it would also be used for Walda's password once we replace the vanilla menu.

VUI widgets: separated construction from [graphics] initialization, so that menu states can be created in advance of setting up BG layers, graphical resources, and so on.

Modified the Makefile so that C files can be nested one folder level deeper.