Window and malloc debugging; short string entry menu work
Added debugging code to the GF window library, to help identify reasons why a window may fail to spawn, and to the GF malloc library, to investigate heap allocation failures. (In this case, we were writing past the end of a window's tile graphics buffer and corrupting information for the next block in the heap.)

Fixed palette issues with the short string entry menu's menu buttons. Turns out, if a PNG and a PAL have the same name and you INCBIN a GBAPAL, the GBAPAL will be synthesized from the PNG instead of converted from the PAL. This is, I presume, a deliberate feature to make it more convenient to use palettized PNGs. I do not like this feature.

Added more explicit size checks to BlitBitmapRect4BitRemapped. These shouldn't be strictly necessary for anything I'm doing right now -- they weren't why we were buffer-overflowing -- but I feel it's safest this way.

Fixed charset button press handlers for the short string entry menu.

In general, the short string entry menu's refactor and redesign are almost complete.