Short string entry menu work
The short string entry menu is now used for renaming PC boxes, for entering your name when starting a new playthrough, and for entering Walda's password.

Extended VUIKeyboardValue to hold 16 characters, which is sufficient for Walda's 15-character passwords. Adjusted the short string entry menu to move the value's on-screen position when the max value length is long; the code for this is a clumsy hack, but it works for now.