WIP HiColor sprite system (ROM does not yet build)
Started work on a system for HiColor sprites, taking advantage of an h-blank interrupt. The system is unfinished, with _Static_assert declarations used to mark to-do notes, so the ROM cannot build until I finish the system (or until these files are excluded from compilation somehow).

Added files with some helper functions, such as analogues to some of C++'s <bit> functions, and insertion sort implementations. The former are mostly based on the famous "bit twiddling hacks" page. It's worth noting that the "block sort" function is incomplete and contains a static assertion, but since it's a header file, you can just... not #include it. I'll probably delete it later; this commit is just a "make sure I don't lose anything if my PC gets annihilated by a thunderstorm or something" commit. I'm committing directly to main because this entire repo is my "dev" branch; the project ain't done yet.