HiColor: optimizations; attempting hand-written assembly
We now track the live buffer instead of the staging buffer, so the h-blank operation doesn't need to XOR the variable.

Disabled some more of the non-essential logic in the update code.

Started experimenting with handwritten assembly for the h-blank handler. It currently works, but I don't think it's optimal yet.