HiColor ARM assembly -- success!
ARM executes fast enough, even from ROM, to handle the bulk of the h-blank logic and data copying on time. It probably helps a lot that we can read the per-scanline palette mappings four at a time (i.e. reading a whole DWORD and shifting it for bytes), since we have more registers to work with.

With the ARM code, we no longer need PALETTES_EARLY_BY_SCANLINES. I'll leave it as a configuration option, but I can probably remove some of the other bandaids e.g. LOST_CAUSE_SCANLINES.

The code is still disorganized. We should have a file with common macros, so we can centrally configure both the assembly files and the C files via those macros. (Each assembly file has its own constant to indicate whether we need the ROM-to-IWRAM trampoline; and right now, the handwritten THUMB and ARM both make it into the ROM, whereas if we could gate them behind the same macro as in the C file, we could have just the one we need in the final game.)

HiColor as a system also still needs some work. For example, the game's palette-fade functions don't touch the HiColor buffers, so HiColor sprites wouldn't fade when all other sprites do; HiColor needs to offer palette functions that the game's other palette functions can call (and the offered APIs need to early-out ASAP if there are no active HiColor palettes).